// recursion-ai/src/app/core/api/model/zidiumWebServiceFrontCategoryListDto.model.ts
/**
 * Zidium Front
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ZidiumWebServiceFrontCategoryDto } from './zidiumWebServiceFrontCategoryDto.model';


export interface ZidiumWebServiceFrontCategoryListDto { 
    items: Array<ZidiumWebServiceFrontCategoryDto>;
    canAdd: boolean;
}